# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: Know API
  description: |
    REST API for the **Know** knowledge management server — a persistent knowledge layer for AI agents backed by SurrealDB.

    ## Quick Start

    1. **Get a token** — run `know db seed` to bootstrap the database and receive a `kh_…` API token, or use the [OIDC device flow](#tag/Auth) for browser-based login.
    2. **Authenticate** — pass the token as a Bearer header:
       ```
       Authorization: Bearer kh_abc123…
       ```
    3. **Create content** — [upsert documents](#tag/Documents/POST/vaults/{vault}/documents) into a vault. Frontmatter (`---\nlabels: [go, api]\n---`) is parsed automatically.
    4. **Search** — [hybrid search](#tag/Search) combines keyword and semantic matching for relevant results.
    5. **Chat with the agent** — [start a chat](#tag/Agent/POST/vaults/{vault}/agent/chat), then subscribe to the [SSE event stream](#tag/Agent/GET/agent/events/{id}) for streamed responses, tool calls, and approval interrupts.

    ## Common Patterns

    | Pattern | Detail |
    |---------|--------|
    | **List envelope** | All list endpoints return `{ "items": [...], "total": N }`. |
    | **Vault name** | Vault-scoped paths use `{vault}` as a **name** (e.g. `default`), resolved server-side to the vault ID. |
    | **Error format** | Errors follow [RFC 9457 Problem Details](https://www.rfc-editor.org/rfc/rfc9457) with `application/problem+json` content type. |
...

    ## Server Layout

...
  version: 0.1.0
servers:
...
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "Token from `know db seed`. Format: `kh_...`"
  parameters:
    vault:
      name: vault
      in: path
      required: true
      description: Vault name (e.g. `default`)
      schema:
        type: string
...
paths:
...
  /vaults:
    get:
      tags: [Vaults]
      summary: List vaults
      description: Returns vaults accessible to the authenticated user, including federated remote vaults.
      responses:
        "200":
          description: List of vaults
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Vault"
                  total:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
  /vaults/{vault}:
    get:
      tags: [Vaults]
      summary: Get vault info
      description: Returns comprehensive stats about a vault (document counts, chunk stats, labels, members, etc.).
      parameters:
        - $ref: "#/components/parameters/vault"
      responses:
        "200":
          description: Vault info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultInfo"
        "404":
          $ref: "#/components/responses/NotFound"
  /vaults/{vault}/settings:
    get:
      tags: [Vaults]
      summary: Get vault settings
      parameters:
        - $ref: "#/components/parameters/vault"
      responses:
        "200":
          description: Current vault settings (with defaults applied)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultSettings"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Vaults]
      summary: Update vault settings
      description: Partially updates vault settings. Only provided fields are changed. Requires admin role.
      parameters:
        - $ref: "#/components/parameters/vault"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultSettings"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultSettings"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ── Documents ───────────────────────────────────────────
  /vaults/{vault}/documents:
    get:
      tags: [Documents]
      summary: Get document by path
      parameters:
        - $ref: "#/components/parameters/vault"
        - name: path
          in: query
          required: true
          schema:
            type: string
          description: Document path (e.g. `/notes/hello.md`)
      responses:
        "200":
          description: Document with content and resolved wiki-links
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Documents]
      summary: Create or update a document
      parameters:
        - $ref: "#/components/parameters/vault"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertDocumentRequest"
      responses:
        "200":
          description: Upserted document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Documents]
      summary: Delete document or folder
      description: |
        Deletes a document or folder. For folders, `recursive=true` is required.
        Use `dry-run=true` to preview what would be deleted.
      parameters:
        - $ref: "#/components/parameters/vault"
        - name: path
          in: query
          required: true
          schema:
            type: string
        - name: recursive
          in: query
          schema:
            type: boolean
        - name: dry-run
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Deletion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteDocumentsResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /vaults/{vault}/documents/ls:
    get:
      tags: [Documents]
      summary: List directory contents
      parameters:
        - $ref: "#/components/parameters/vault"
        - name: path
          in: query
          schema:
            type: string
            default: "/"
          description: Folder path to list
        - name: recursive
          in: query
          schema:
            type: boolean
        - name: labels
          in: query
          schema:
            type: string
          description: Comma-separated label filter. When set, only files matching all labels are returned (with Labels and Source fields populated).
      responses:
        "200":
          description: Directory entries (files and folders)
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/FileEntry"
                  total:
                    type: integer
  /vaults/{vault}/documents/move:
    post:
      tags: [Documents]
      summary: Move a document or folder
      parameters:
        - $ref: "#/components/parameters/vault"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveRequest"
      responses:
        "200":
          description: Move result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoveResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Destination already exists
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /vaults/{vault}/documents/bulk:
    post:
      tags: [Documents]
      summary: Bulk upload documents and assets
      description: |
        Multipart upload. First part must be `meta` (JSON with `force` and `dryRun` flags).
        Subsequent parts are files identified by form name (vault path).
      parameters:
        - $ref: "#/components/parameters/vault"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                meta:
                  type: string
                  description: 'JSON: `{"force": false, "dryRun": false}`'
      responses:
        "200":
          description: Per-file results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkResponse"
  /vaults/{vault}/documents/clip:
    post:
      tags: [Documents]
      summary: Fetch a web page and save as document
      description: Uses Jina Reader to convert a URL to markdown and saves it in the vault.
      parameters:
        - $ref: "#/components/parameters/vault"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FetchRequest"
      responses:
        "200":
          description: Saved document info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FetchResponse"

  # ── Import ──────────────────────────────────────────────
  /vaults/{vault}/import/manifest:
    post:
      tags: [Import]
      summary: "Phase 1: Check which files need uploading"
      description: |
        Send a list of files with their SHA256 hashes. The server responds with
        which files need to be uploaded (phase 2) and which can be skipped.
      parameters:
        - $ref: "#/components/parameters/vault"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportManifestRequest"
      responses:
        "200":
          description: Manifest check results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportManifestResponse"
  /vaults/{vault}/import/upload:
    post:
      tags: [Import]
      summary: "Phase 2: Upload needed files"
      description: |
        Multipart upload of files indicated as "needed" by the manifest phase.
        First part must be `meta` (JSON with `hashes` map: path → SHA256).
        Subsequent parts are files identified by form name (vault path).
        Hash verification is enforced — mismatches abort the import.
      parameters:
        - $ref: "#/components/parameters/vault"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                meta:
                  type: string
                  description: 'JSON: `{"hashes": {"path": "sha256hash"}}`'
      responses:
        "200":
          description: Per-file upload results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportUploadResponse"
        "400":
          description: Hash mismatch or oversized file
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"

  # ── Export ──────────────────────────────────────────────
  /vaults/{vault}/export:
    get:
      tags: [Export]
      summary: Export vault as tar.gz
      description: Streams the vault contents as a gzipped tar archive.
      parameters:
        - $ref: "#/components/parameters/vault"
      responses:
        "200":
          description: Vault archive
          content:
            application/gzip:
              schema:
                type: string
                format: binary
  /vaults/{vault}/export/epub:
    get:
      tags: [Export]
      summary: Export vault as EPUB
      description: |
        Exports a single document or all markdown files under a folder as an EPUB book.
        When `path` points to a file, that file becomes the sole chapter.
        When `path` points to a folder, all `.md` files underneath are included.
      parameters:
        - $ref: "#/components/parameters/vault"
        - name: path
          in: query
          required: true
          schema:
            type: string
          description: Document or folder path to export
        - name: title
          in: query
          schema:
            type: string
          description: EPUB title (defaults to document title or folder name)
        - name: author
          in: query
          schema:
            type: string
          description: EPUB author metadata
      responses:
        "200":
          description: EPUB file
          content:
            application/epub+zip:
              schema:
                type: string
                format: binary

  # ── Backup / Restore ────────────────────────────────────
  /vaults/{vault}/backup:
    get:
      tags: [Backup]
      summary: Export vault backup
      description: Manifest-based backup as tar.gz archive including all documents, assets, and metadata.
      parameters:
        - $ref: "#/components/parameters/vault"
      responses:
        "200":
          description: Backup archive
          content:
            application/gzip:
              schema:
                type: string
                format: binary
  /backup/restore:
    post:
      tags: [Backup]
      summary: Restore vault from backup
      description: |
        Restores a vault from a tar.gz backup archive. Creates the vault if it doesn't exist.
        Send the archive as the raw request body with `Content-Type: application/gzip`.
      requestBody:
        required: true
        content:
          application/gzip:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Restore completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string

  # ── Search ──────────────────────────────────────────────
  /vaults/{vault}/search:
    get:
      tags: [Search]
      summary: Search documents
      description: Hybrid BM25 + vector search with reciprocal rank fusion.
      parameters:
        - $ref: "#/components/parameters/vault"
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: labels
          in: query
          schema:
            type: string
          description: Comma-separated label filter
        - name: bm25_only
          in: query
          schema:
            type: boolean
            default: false
          description: When true, use BM25 keyword search only (skip vector search)
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/SearchResult"
                  total:
                    type: integer

  # ── Folders ─────────────────────────────────────────────
  /vaults/{vault}/folders:
    get:
      tags: [Folders]
      summary: List folders
      parameters:
        - $ref: "#/components/parameters/vault"
        - name: parent
          in: query
          schema:
            type: string
          description: Filter to immediate children of this folder
      responses:
        "200":
          description: List of folders
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/FolderResponse"
                  total:
                    type: integer
    patch:
      tags: [Folders]
      summary: Update folder settings
      description: Currently supports toggling the `no_embed` flag. When enabled, strips existing embeddings.
      parameters:
        - $ref: "#/components/parameters/vault"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFolderRequest"
      responses:
        "200":
          description: Update result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateFolderResponse"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Versions ────────────────────────────────────────────
  /vaults/{vault}/versions:
    get:
      tags: [Versions]
      summary: List document versions
      parameters:
        - $ref: "#/components/parameters/vault"
        - name: path
          in: query
          required: true
          schema:
            type: string
          description: Document path
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Version history
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/VersionResponse"
                  total:
                    type: integer
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Labels ──────────────────────────────────────────────
  /vaults/{vault}/documents/labels:
    patch:
      tags: [Labels]
      summary: Add or remove labels on a document
      parameters:
        - $ref: "#/components/parameters/vault"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
                  description: Document path
                add:
                  type: array
                  items:
                    type: string
                  description: Labels to add
                remove:
                  type: array
                  items:
                    type: string
                  description: Labels to remove
      responses:
        "200":
          description: Updated labels
          content:
            application/json:
              schema:
                type: object
                properties:
                  labels:
                    type: array
                    items:
                      type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /vaults/{vault}/labels:
    get:
      tags: [Labels]
      summary: List labels
      description: Returns label names, or label names with document counts when `counts=true`.
      parameters:
        - $ref: "#/components/parameters/vault"
        - name: counts
          in: query
          schema:
            type: boolean
          description: Include document counts per label
      responses:
        "200":
          description: Label list (string array) or label counts (object array), wrapped in items/total envelope
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    oneOf:
                      - type: array
                        items:
                          type: string
                      - type: array
                        items:
                          $ref: "#/components/schemas/LabelCount"
                  total:
                    type: integer

  # ── Tasks ───────────────────────────────────────────────
  /vaults/{vault}/bookmarks:
    get:
      tags: [Bookmarks]
      summary: List bookmarks
      description: Lists the current user's bookmarked files and folders in the vault.
      parameters:
        - $ref: "#/components/parameters/vault"
      responses:
        "200":
          description: Bookmarked files
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/FileEntry"
                  total:
                    type: integer
    put:
      tags: [Bookmarks]
      summary: Add bookmark
      description: Bookmark a file or folder by path. Idempotent.
      parameters:
        - $ref: "#/components/parameters/vault"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
                  description: Vault-relative path to bookmark
                  example: "/notes/important.md"
      responses:
        "204":
          description: Bookmark added
    delete:
      tags: [Bookmarks]
      summary: Remove bookmark
      description: Remove a bookmark by file path. Idempotent.
      parameters:
        - $ref: "#/components/parameters/vault"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
                  description: Vault-relative path to unbookmark
      responses:
        "204":
          description: Bookmark removed
  /vaults/{vault}/tasks:
    get:
      tags: [Tasks]
      summary: List tasks
      description: Lists tasks extracted from document checkboxes, with optional filters.
      parameters:
        - $ref: "#/components/parameters/vault"
        - name: status
          in: query
          schema:
            type: string
            enum: [open, done]
        - name: labels
          in: query
          schema:
            type: string
          description: Comma-separated label filter
        - name: due_before
          in: query
          schema:
            type: string
            format: date
          description: "Filter: due date before (YYYY-MM-DD)"
        - name: due_after
          in: query
          schema:
            type: string
            format: date
          description: "Filter: due date after (YYYY-MM-DD)"
        - name: folder
          in: query
          schema:
            type: string
        - name: path
          in: query
          schema:
            type: string
          description: Filter to tasks from a specific document
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated task list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListResponse"
  /vaults/{vault}/tasks/{id}/toggle:
    post:
      tags: [Tasks]
      summary: Toggle task status
      description: Toggles a task between `open` and `done`, updating the source document's checkbox.
      parameters:
        - $ref: "#/components/parameters/vault"
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"

  # ── Assets ──────────────────────────────────────────────
  /vaults/{vault}/assets:
    post:
      tags: [Assets]
      summary: Upload an image asset
      parameters:
        - $ref: "#/components/parameters/vault"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [path, file]
              properties:
                path:
                  type: string
                  description: Vault path for the asset
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Uploaded asset metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetMeta"
    get:
      tags: [Assets]
      summary: Download asset binary
      parameters:
        - $ref: "#/components/parameters/vault"
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Asset binary data
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Assets]
      summary: Delete an asset
      parameters:
        - $ref: "#/components/parameters/vault"
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Asset deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /vaults/{vault}/assets/meta:
    get:
      tags: [Assets]
      summary: Get asset metadata
      parameters:
        - $ref: "#/components/parameters/vault"
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Asset metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetMeta"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── External Links ──────────────────────────────────────
  /vaults/{vault}/external-links:
    get:
      tags: [External Links]
      summary: List external links
      parameters:
        - $ref: "#/components/parameters/vault"
        - name: hostname
          in: query
          schema:
            type: string
        - name: file
          in: query
          schema:
            type: string
          description: Filter by file ID
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated link list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExternalLinkListResponse"
  /vaults/{vault}/external-links/stats:
    get:
      tags: [External Links]
      summary: External link stats by hostname
      parameters:
        - $ref: "#/components/parameters/vault"
      responses:
        "200":
          description: Link count per hostname
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExternalLinkStatsResponse"

  # ── Changes (incremental sync) ──────────────────────────
...
  /agent/events/{id}:
    get:
      tags: [Agent]
      summary: SSE stream for agent response
      description: |
        Server-Sent Events stream for a running agent conversation.
        Events include text chunks, tool calls, tool results, and completion.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Stream ID from the chat response
      responses:
        "200":
          description: SSE event stream
          content:
            text/event-stream:
              schema:
                type: string
  /agent/cancel/{id}:
    post:
      tags: [Agent]
      summary: Cancel an agent run
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cancellation requested
...